requirements.txt
black==21.8b0
click==8.0.1
colorama==0.4.4
isort==5.9.3
loguru==0.5.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
python-Levenshtein==0.12.2
regex==2021.8.28
tomli==1.2.1
typing-extensions==3.10.0.2
win32-setctime==1.0.3
